Add Google login endpoint for iOS
