+++ requirements.txt
@@ -1,6 +1,7 @@
 aspy.yaml==1.3.0
 beautifulsoup4==4.8.1
 bs4==0.0.1
+cachetools==4.0.0
 certifi==2019.9.11
 cfgv==2.0.1
 chardet==3.0.4
@@ -9,6 +10,7 @@ entrypoints==0.3
 flake8==3.7.9
 Flask==1.1.1
 Flask-SQLAlchemy==2.4.1
+google-auth==1.11.0
 identify==1.4.7
 idna==2.8
 importlib-metadata==0.23
@@ -20,10 +22,13 @@ mccabe==0.6.1
 more-itertools==7.2.0
 nodeenv==1.3.3
 pre-commit==1.20.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
 pycodestyle==2.5.0
 pyflakes==2.1.1
 PyYAML==5.1.2
 requests==2.22.0
+rsa==4.0
 six==1.13.0
 soupsieve==1.9.5
 SQLAlchemy==1.3.11
