+++ src/app/coursegrab/__init__.py
@@ -1,8 +1,6 @@
 from flask import Blueprint
-from app.coursegrab.controllers.create_user_controller import *
 from app.coursegrab.controllers.hello_world_controller import *
-from app.coursegrab.controllers.login_controller import *
-from app.coursegrab.controllers.oauth2_callback_controller import *
+from app.coursegrab.controllers.initialize_session_controller import *
 from app.coursegrab.controllers.retrieve_tracking_controller import *
 from app.coursegrab.controllers.track_course_controller import *
 from app.coursegrab.controllers.untrack_course_controller import *
@@ -12,10 +10,8 @@ from app.coursegrab.controllers.update_session_controller import *
 coursegrab = Blueprint("coursegrab", __name__, url_prefix="/api")
 
 controllers = [
-    CreateUserController(),
     HelloWorldController(),
-    LoginController(),
-    OAuth2CallbackController(),
+    InitializeSessionController(),
     RetrieveTrackingController(),
     TrackCourseController(),
     UntrackCourseController(),
